Implement Readers for LogicalGroup-Based Replication.
