+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/logreader/LogicalGroupSnapshotReader.java
@@ -3,23 +3,19 @@ package org.corfudb.infrastructure.logreplication.replication.send.logreader;
 import org.corfudb.infrastructure.logreplication.infrastructure.LogReplicationContext;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.LogReplication.LogReplicationSession;
-import java.util.UUID;
 
 /**
  * Snapshot reader implementation for Logical Grouping Replication Model.
  */
-public class LogicalGroupSnapshotReader extends SnapshotReader {
+public class LogicalGroupSnapshotReader extends BaseSnapshotReader {
 
     public LogicalGroupSnapshotReader(CorfuRuntime runtime, LogReplicationSession session,
                                       LogReplicationContext replicationContext) {
+        super(runtime, session, replicationContext);
     }
 
     @Override
-    public SnapshotReadMessage read(UUID syncRequestId) {
-        return null;
-    }
+    protected void refreshStreamsToReplicateSet() {
 
-    @Override
-    public void reset(long ts) {
     }
 }
