+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/logreader/RoutingQueuesSnapshotReader.java
@@ -8,15 +8,17 @@ import java.util.UUID;
 /**
  * Snapshot reader implementation for Routing Queues Replication Model.
  */
-public class RoutingQueuesSnapshotReader extends SnapshotReader {
+public class RoutingQueuesSnapshotReader extends BaseSnapshotReader {
 
     public RoutingQueuesSnapshotReader(CorfuRuntime corfuRuntime, LogReplicationSession session,
                                        LogReplicationContext replicationContext) {
+        super(corfuRuntime, session, replicationContext);
     }
 
     @Override
-    public SnapshotReadMessage read(UUID syncRequestId) {
-        return null;
+    protected void refreshStreamsToReplicateSet() {
+        throw new IllegalStateException("Unexpected workflow encountered.  Stream UUIDs cannot be refreshed for this " +
+            "model");
     }
 
     @Override
